pyproject.toml
[tool.poetry]
name = "token-validation"
version = "0.1.0"
description = ""
authors = ["Ethan-ahnlabio <ethan@ahnlab.io>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.0.3"
web3 = "^6.6.1"
slither-analyzer = "^0.9.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
